Make zlocal per thread
Currently zlocal is computed once.

Thus if blockassembler is called in an openmp environment
the same zlocal is used by all threads

Now an array of zlocals is established (its length  equal to #threads)
to prevent this.

NOTE: When multithreading is active, the unit test test_nitsche.jl
fails. This has been true already before this change and should thus be
fixed separately.