Adapt SauterSchwab and Wilton ints for HH3D
Added functions so that all combinations of patch and pyramid functions can be used
with Sauter integrals for the common face/edge/certex case and a singularity extraction
rule for the near singular case.

Not clear why DoubleQuadStrat performs better for common vertex than
singularity extraction. For triangles close together singularity extraction
appears to perform better as is expected.