Add Lippmann-Schwinger VIE
We add Lippmann-Schwinger VIE in two flavors

- with one partial integration
  (hhvolume & hhboundary functions in vie.jl)
- no partial integration (one derivative on the Green's function)
  (hhvolumegradG)

In addition, the function hhvolumek0 is needed for k != 0 in the
Helmholtz equation.

Note that more flavors exist, which yet need to be implemented.
For example, the one described in

Bleszynski, E., M. Bleszynski, and T. Jaroszewicz. “Fast Volumetric
Integral-Equation Solver for Acoustic Wave Propagation through
Inhomogeneous Media.” The Journal of the Acoustical Society of America
124, no. 1 (July 1, 2008): 396–408. https://doi.org/10.1121/1.2924203.

where one derivative is moved onto the contrast function.

TODO: Add unit test for accoustic version (once it is available in
SphericalScattering. So far only the electrostatic version is
tested)