Add baseline concept of FEMFunctions
FEMFunction is a type that combines expansion coefficients and
function space.

Also added is a routine Lp_integrate that integrates FEMFunctions
or linear combinations over their domain.

Moreover, in order to measure errors, analytic functions can be
wrapped in the type GlobalFunction, which can be linearly combined
with FEMFunctions. Thus, the Lp-integral over the difference of a
FEMFunction and an analytic function can be computed to measure, for
example, an error (including interpolation error).

TODO:
 - It might come in handy if the returntype of a GlobalFunction is predicted
 - FEMFunction could be integrated into solve functions etc. Maybe an extension
   for time-domain is, however, necessary.