2D-Helmholtz: Add TangentTrace, ẑ x support, and Mie integration test
Most textbooks formulate the 2D TM-MFIE and the 2D TE-EFIE using
the tangential trace of the magnetic and eletric field, respectively.

A tangent trace is added to allow following textbooks more closely.

Moreover, the potential operators lead to gradient fields instead of
curls. By providing access to ẑ x, those 2D-gradients can be conveniently
rotated into curls.

Lastly, an integration test with respect to the Mie series is included.
It also shows and tests the alpha and beta parameter of the operators
as well as certain right-hand side operations.

Moreover, some cleanup was performed removing unnecessary functions
of the Helmholtz 2D excitations.