Refactor hh2d near and add more excitations
1) Refactoring:
Previously, there was no way to define alpha and beta parameters
as for the actual operators. This made a consistent usage difficult.

Moreover, the hypersingular operator now only has a weak singularity.

2) More Excitations:

TM and TE scattering of a PEC cylinder due to an infinitely long line
current excitation.

Analytical Solutions:

The TM implementation follows Section 6.5.3 Jin Theory and Computation
of Electromagnetic fields.

The TE scattering implementation is derived using a phi-directed
current.

Numerical Solutions:

The operators used are the same as in test_hh2d_mie.jl where a planewave
excitation was used.