Add 1D Higher Order Lagrange Functions (Cx and C0)
Polynomials are implemented hard-coded up to order 10
for maximal speed.

TODO: The Mie unit test should be adapted once
curvilinear meshes are available. So far, we only
confirm that higher-order functions do not break the
performance. However, since the mesh is not refined
the geometrical error starts to dominate quickly.