Example implementation of 1D curvilinear elements
Requires an updated CMS version!

Some explanations:

1) Overall, the changes to the existing code are minimal.
My understanding is that the way Sauter-Schwab works these days
does not necessitate a swapping of the (control) vertices,
instead we work with a reference element and basically remap the
barycentric coordinates.

I adapted the unit test for 1D Sauter-Schwab. I do not understand
why I picked SauterSchwab as a reference for itself. Now I use
the Gauss-Legendre. Of course, accuracy is low, but if something would
go fundamentally wrong with the normal or the reordering of vertices
I would expect it to be visible.

2) My 2D Mie series suffered from early convergence thus the rewrite.

3) The changes for lagrange.jl where necessary because the number of vertices
of an element does not help to assess the type in the higher order case.
(3 vertices can be a triangle or a quadratic segment)